move sb_not_here() and related function to a new file
