+++ libs/libarchfpga/src/switchblock_types.h
@@ -169,6 +169,15 @@ struct t_wireconn_inf {
 /* Use a map to index into the string permutation functions used to connect from one side to another */
 typedef std::map<SBSideConnection, std::vector<std::string>> t_permutation_map;
 
+struct t_specified_loc {
+    int x = -1; ///< The exact x-axis location that this SB is used, meaningful when type is set to E_XY_specified
+    int y = -1; ///< The exact y-axis location that this SB is used, meaningful when type is set to E_XY_specified
+
+    // We can also define a region to apply this SB to all locations falls into this region using regular expression in the architecture file
+    t_sb_loc_spec reg_x;
+    t_sb_loc_spec reg_y;
+};
+
 /**
  * @brief Lists all information about a particular switch block specified in the architecture file
  */
@@ -177,14 +186,9 @@ struct t_switchblock_inf {
     e_sb_location location;          ///< where on the FPGA this switchblock should be built (i.e. perimeter, core, everywhere)
     e_directionality directionality; ///< the directionality of this switchblock (unidir/bidir)
 
-    int x = -1; ///< The exact x-axis location that this SB is used, meaningful when type is set to E_XY_specified
-    int y = -1; ///< The exact y-axis location that this SB is used, meaningful when type is set to E_XY_specified
-
-    /* We can also define a region to apply this SB to all locations falls into this region using regular expression in the architecture file*/
-    t_sb_loc_spec reg_x;
-    t_sb_loc_spec reg_y;
-
     t_permutation_map permutation_map; ///< map holding the permutation functions attributed to this switchblock
+
+    t_specified_loc specified_loc;
 
     std::vector<t_wireconn_inf> wireconns; ///< list of wire types/groups this SB will connect
 };
