+++ vpr/src/route/rr_graph_generation/build_switchblocks.cpp
@@ -140,6 +140,7 @@
 #include "parse_switchblocks.h"
 #include "vtr_expr_eval.h"
 #include "rr_types.h"
+#include "switchblock_scatter_gather_common_utils.h"
 
 using vtr::FormulaParser;
 using vtr::t_formula_data;
@@ -330,88 +331,6 @@ static int get_switchpoint_of_wire(const DeviceGrid& grid,
                                    int seg_coord,
                                    e_side sb_side);
 
-/**
- * @brief check whether a switch block exists in a specified coordinate within the device grid
- *
- *   @param grid device grid
- *   @param inter_cluster_rr used to check whether inter-cluster programmable routing resources exist in the current layer
- *   @param x x-coordinate of the location
- *   @param y y-coordinate of the location
- *   @param layer layer-coordinate of the location
- *   @param sb switchblock information specified in the architecture file
- *
- * @return true if a switch block exists at the specified location, false otherwise.
- */
-static bool sb_not_here(const DeviceGrid& grid,
-                        const std::vector<bool>& inter_cluster_rr,
-                        int x,
-                        int y,
-                        int layer,
-                        const t_switchblock_inf& sb);
-
-/**
- * @brief check whether specified coordinate is located at the device grid corner and a switch block exists there
- *
- *   @param grid device grid
- *   @param inter_cluster_rr used to check whether inter-cluster programmable routing resources exist in the current layer
- *   @param x x-coordinate of the location
- *   @param y y-coordinate of the location
- *   @param layer layer-coordinate of the location
- *
- * @return true if the specified coordinate represents a corner location within the device grid and a switch block exists there, false otherwise.
- */
-static bool is_corner_sb(const DeviceGrid& grid, const std::vector<bool>& inter_cluster_rr, int x, int y, int layer);
-
-/**
- * @brief check whether specified coordinate is located at one of the perimeter device grid locations and a switch block exists there
- *
- *   @param grid device grid
- *   @param inter_cluster_rr used to check whether inter-cluster programmable routing resources exist in the current layer
- *   @param x x-coordinate of the location
- *   @param y y-coordinate of the location
- *   @param layer layer-coordinate of the location
- *
- * @return true if the specified coordinate represents a perimeter location within the device grid and a switch block exists there, false otherwise.
- */
-static bool is_perimeter_sb(const DeviceGrid& grid, const std::vector<bool>& inter_cluster_rr, int x, int y, int layer);
-
-/**
- * @brief check whether specified coordinate is located at core of the device grid (not perimeter) and a switch block exists there
- *
- *   @param grid device grid
- *   @param inter_cluster_rr used to check whether inter-cluster programmable routing resources exist in the current layer
- *   @param x x-coordinate of the location
- *   @param y y-coordinate of the location
- *   @param layer layer-coordinate of the location
- *
- * @return true if the specified coordinate represents a core location within the device grid and a switch block exists there, false otherwise.
- */
-static bool is_core_sb(const DeviceGrid& grid, const std::vector<bool>& inter_cluster_rr, int x, int y, int layer);
-
-/**
- * @brief check whether specified coordinate is located in the architecture-specified regions that the switchblock should be applied to
- *
- *   @param grid device grid
- *   @param inter_cluster_rr used to check whether inter-cluster programmable routing resources exist in the current layer
- *   @param x x-coordinate of the location
- *   @param y y-coordinate of the location
- *   @param sb switchblock information specified in the architecture file
- *
- * @return true if the specified coordinate falls into the architecture-specified location for this switchblock, false otherwise.
- */
-static bool match_sb_xy(const DeviceGrid& grid, const std::vector<bool>& inter_cluster_rr, int x, int y, int layer, const t_switchblock_inf& sb);
-
-/**
- * @brief check whether specified layer has inter-cluster programmable routing resources or not.
- *
- *   @param grid device grid
- *   @param inter_cluster_rr inter-cluster programmable routing resources availability within different layers in multi-die FPGAs
- *   @param layer a valid layer index within the device grid, must be between [0..num_layer-1]
- *
- * @return true if the specified layer contain inter-cluster programmable routing resources, false otherwise.
- */
-static bool is_prog_routing_avail(const DeviceGrid& grid, const std::vector<bool>& inter_cluster_rr, int layer);
-
 /* adjusts a negative destination wire index calculated from a permutation formula */
 static int adjust_formula_result(int dest_wire, int src_W, int dest_W, int connection_ind);
 
@@ -452,11 +371,11 @@ t_sb_connection_map* alloc_and_load_switchblock_permutations(const t_chan_detail
             VPR_FATAL_ERROR(VPR_ERROR_ARCH, "alloc_and_load_switchblock_connections: Switchblock %s does not match directionality of architecture\n", sb.name.c_str());
         }
 
-        // Iterate over the x,y, layer coordinates spanning the FPGA, filling in all the switch blocks that exist
+        // Iterate over all coordinates spanning the FPGA, filling in all the switch blocks that exist
         for (int layer_coord = 0; layer_coord < grid.get_num_layers(); layer_coord++) {
             for (size_t x_coord = 0; x_coord < grid.width(); x_coord++) {
                 for (size_t y_coord = 0; y_coord <= grid.height(); y_coord++) {
-                    if (sb_not_here(grid, inter_cluster_rr, x_coord, y_coord, layer_coord, sb)) {
+                    if (sb_not_here(grid, inter_cluster_rr, x_coord, y_coord, layer_coord, sb.location, sb.specified_loc)) {
                         continue;
                     }
 
@@ -489,161 +408,6 @@ void free_switchblock_permutations(t_sb_connection_map* sb_conns) {
      * this significantly reduces memory usage during the routing stage when running multiple
      * large benchmark circuits in parallel. */
     vtr::malloc_trim(0);
-}
-
-static bool sb_not_here(const DeviceGrid& grid,
-                        const std::vector<bool>& inter_cluster_rr,
-                        int x,
-                        int y,
-                        int layer,
-                        const t_switchblock_inf& sb) {
-    bool sb_not_here = true;
-
-    switch (sb.location) {
-        case e_sb_location::E_EVERYWHERE:
-            sb_not_here = false;
-            break;
-        case e_sb_location::E_PERIMETER:
-            if (is_perimeter_sb(grid, inter_cluster_rr, x, y, layer)) {
-                sb_not_here = false;
-            }
-            break;
-        case e_sb_location::E_CORNER:
-            if (is_corner_sb(grid, inter_cluster_rr, x, y, layer)) {
-                sb_not_here = false;
-            }
-            break;
-        case e_sb_location::E_CORE:
-            if (is_core_sb(grid, inter_cluster_rr, x, y, layer)) {
-                sb_not_here = false;
-            }
-            break;
-        case e_sb_location::E_FRINGE:
-            if (is_perimeter_sb(grid, inter_cluster_rr, x, y, layer) && !is_corner_sb(grid, inter_cluster_rr, x, y, layer)) {
-                sb_not_here = false;
-            }
-            break;
-        case e_sb_location::E_XY_SPECIFIED:
-            if (match_sb_xy(grid, inter_cluster_rr, x, y, layer, sb)) {
-                sb_not_here = false;
-            }
-
-            break;
-        default:
-            VPR_FATAL_ERROR(VPR_ERROR_ARCH, "sb_not_here: unrecognized location enum: %d\n", sb.location);
-            break;
-    }
-    return sb_not_here;
-}
-
-static bool is_corner_sb(const DeviceGrid& grid, const std::vector<bool>& inter_cluster_rr, int x, int y, int layer) {
-    if (!is_prog_routing_avail(grid, inter_cluster_rr, layer)) {
-        return false;
-    }
-    bool is_corner = false;
-    if ((x == 0 && y == 0) || (x == 0 && y == int(grid.height()) - 2) || //-2 for no perim channels
-        (x == int(grid.width()) - 2 && y == 0) ||                        //-2 for no perim channels
-        (x == int(grid.width()) - 2 && y == int(grid.height()) - 2)) {   //-2 for no perim channels
-        is_corner = true;
-    }
-    return is_corner;
-}
-
-static bool is_perimeter_sb(const DeviceGrid& grid, const std::vector<bool>& inter_cluster_rr, int x, int y, int layer) {
-    if (!is_prog_routing_avail(grid, inter_cluster_rr, layer)) {
-        return false;
-    }
-    bool is_perimeter = false;
-    if (x == 0 || x == int(grid.width()) - 2 || y == 0 || y == int(grid.height()) - 2) {
-        is_perimeter = true;
-    }
-    return is_perimeter;
-}
-
-static bool is_core_sb(const DeviceGrid& grid, const std::vector<bool>& inter_cluster_rr, int x, int y, int layer) {
-    if (!is_prog_routing_avail(grid, inter_cluster_rr, layer)) {
-        return false;
-    }
-    bool is_core = !is_perimeter_sb(grid, inter_cluster_rr, x, y, layer);
-    return is_core;
-}
-
-static bool is_prog_routing_avail(const DeviceGrid& grid, const std::vector<bool>& inter_cluster_rr, int layer) {
-    bool is_prog_avail = true;
-    //make sure layer number is legal
-    VTR_ASSERT(layer >= 0 && layer < grid.get_num_layers());
-    //check if the current layer has programmable routing resources before trying to build a custom switch blocks
-    if (!inter_cluster_rr.at(layer)) {
-        is_prog_avail = false;
-    }
-    return is_prog_avail;
-}
-
-static bool match_sb_xy(const DeviceGrid& grid, const std::vector<bool>& inter_cluster_rr, int x, int y, int layer, const t_switchblock_inf& sb) {
-    if (!is_prog_routing_avail(grid, inter_cluster_rr, layer)) {
-        return false;
-    }
-    //if one of sb_x and sb_y is defined, we either know the exact location (x,y) or the exact x location (will apply it to all rows)
-    //or the exact y location (will apply it to all columns)
-    if (sb.x != -1 || sb.y != -1) {
-        if (x == sb.x && y == sb.y) {
-            return true;
-        }
-
-        if (x == sb.x && sb.y == -1) {
-            return true;
-        }
-
-        if (sb.x == -1 && y == sb.y) {
-            return true;
-        }
-    }
-
-    //if both sb_x and sb_y is not defined, we have a region that we should apply this SB pattern to, we just need to check
-    //whether the location passed into this function falls within this region or not
-    //calculate the appropriate region based on the repeatx/repeaty and current location.
-    //This is to determine whether the given location is part of the current SB specified region with regular expression or not
-    //After region calculation, the current SB will apply to this location if:
-    // 1) the given (x,y) location falls into the calculated region
-    // *AND*
-    // 2) incrx/incry are respected within the region, this means all locations within the calculated region do
-    //    not necessarily crosspond to the current SB. If incrx/incry is equal to 1, then all locations within the
-    //    calculated region are valid.
-
-    //calculate the region
-    int x_reg_step = (sb.reg_x.repeat != 0) ? (x - sb.reg_x.start) / sb.reg_x.repeat : sb.reg_x.start;
-    int y_reg_step = (sb.reg_y.repeat != 0) ? (y - sb.reg_y.start) / sb.reg_y.repeat : sb.reg_y.start;
-
-    //step must be non-negative
-    x_reg_step = std::max(0, x_reg_step);
-    y_reg_step = std::max(0, y_reg_step);
-
-    int reg_startx = sb.reg_x.start + (x_reg_step * sb.reg_x.repeat);
-    int reg_endx = sb.reg_x.end + (x_reg_step * sb.reg_x.repeat);
-    reg_endx = std::min(reg_endx, int(grid.width() - 1));
-
-    int reg_starty = sb.reg_y.start + (y_reg_step * sb.reg_y.repeat);
-    int reg_endy = sb.reg_y.end + (y_reg_step * sb.reg_y.repeat);
-    reg_endy = std::min(reg_endy, int(grid.height() - 1));
-
-    //check x coordinate
-    if (x >= reg_startx && x <= reg_endx) { //should fall into the region
-        //we also should respect the incrx
-        //if incrx is not equal to 1, all locations within this region are *NOT* valid
-        if ((x + reg_startx) % sb.reg_x.incr == 0) {
-            //valid x coordinate, check for y value
-            if (y >= reg_starty && y <= reg_endy) {
-                //check for incry, similar as incrx
-                if ((y + reg_starty) % sb.reg_y.incr == 0) {
-                    //both x and y are valid
-                    return true;
-                }
-            }
-        }
-    }
-
-    //if reach here, we don't have sb in this location
-    return false;
 }
 
 static t_wire_type_sizes count_wire_type_sizes(const t_chan_seg_details* channel, int nodes_per_chan) {
